backend/requirements.txt
Package                       Version  
----------------------------- ---------
asgiref                       3.3.1    
astroid                       2.4.2    
autopep8                      1.5.4    
beautifulsoup4                4.9.3    
certifi                       2020.11.8
cffi                          1.14.4   
chardet                       3.0.4    
coreapi                       2.3.3    
coreschema                    0.0.4    
cryptography                  3.2.1    
defusedxml                    0.7.0rc1 
Django                        3.1.4    
django-cors-headers           3.5.0    
django-environ                0.4.5    
django-templated-mail         1.1.1    
djangorestframework           3.12.2   
djangorestframework-simplejwt 4.6.0    
djoser                        2.1.0    
idna                          2.10     
importlib-metadata            1.7.0    
isort                         5.6.4    
itypes                        1.2.0    
Jinja2                        2.11.2   
lazy-object-proxy             1.4.3    
MarkupSafe                    1.1.1    
mccabe                        0.6.1    
oauthlib                      3.1.0    
Pillow                        8.0.1    
pip                           19.2.3   
psycopg2                      2.8.6    
psycopg2-binary               2.8.6    
pycodestyle                   2.6.0    
pycparser                     2.20     
PyJWT                         1.7.1    
pylint                        2.6.0    
python3-openid                3.2.0    
pytz                          2020.4   
requests                      2.25.0   
requests-oauthlib             1.3.0    
setuptools                    41.2.0   
six                           1.15.0   
social-auth-app-django        4.0.0    
social-auth-core              3.3.3    
soupsieve                     2.1      
sqlparse                      0.4.1    
toml                          0.10.2   
typed-ast                     1.4.1    
uritemplate                   3.0.1    
urllib3                       1.26.2   
wrapt                         1.12.1   
zipp                          3.4.0    
